@@ -10,11 +10,10 @@
   ],
   "license": "MIT",
   "dependencies": {
+    "@vue/compat": "^3.2.26",
     "axios": "^0.21.1",
-    "vue3-runtime-template": "^1.0.2",
-    "vuex": "^3.0.1",
-    "vue": "^3.1.0",
-    "@vue/compat": "^3.1.0"
+    "mitt": "^3.0.0",
+    "vue": "^3.2.26"
   },
   "exports": {
     ".": "./lib/matestack/ui/vue_js/index.js"
