@@ -1,11 +1,11 @@
-import Vue from 'vue'
 import axios from 'axios'
-import VRuntimeTemplate from "vue3-runtime-template"
 import matestackEventHub from '../event_hub'
 import componentMixin from './mixin'
+import componentHelpers from './helpers'
 
 const componentDef = {
   mixins: [componentMixin],
+  template: componentHelpers.inlineTemplate,
   data: function(){
     return {
       asyncTemplate: null,
@@ -54,7 +54,7 @@ const componentDef = {
         method: "get",
         url: location.pathname + location.search,
         headers: {
-          'X-CSRF-Token': document.getElementsByName("csrf-token")[0].getAttribute('content')
+          'X-CSRF-Token': self.getXcsrfToken()
         },
         params: {
           "component_key": self.props["component_key"],
@@ -93,18 +93,13 @@ const componentDef = {
       this.showing = true
     }
   },
-  beforeDestroy: function() {
+  beforeUnmount: function() {
     const self = this
     clearTimeout(self.hideAfterTimeout)
     self.removeEvents(this.props["show_on"], self.show)
     self.removeEvents(this.props["hide_on"], self.hide)
     self.removeEvents(this.props["rerender_on"], self.rerender)
-  },
-  components: {
-    VRuntimeTemplate: VRuntimeTemplate
   }
 }
-
-let component = Vue.component('matestack-ui-core-async', componentDef)
 
 export default componentDef
