@@ -1,10 +1,10 @@
-import Vue from 'vue'
-import VRuntimeTemplate from "vue3-runtime-template"
 import matestackEventHub from '../event_hub'
 import componentMixin from './mixin'
+import componentHelpers from './helpers'
 
 const componentDef = {
   mixins: [componentMixin],
+  template: componentHelpers.inlineTemplate,
   props: {
     initialTemplate: String,
   },
@@ -81,7 +81,7 @@ const componentDef = {
     this.registerEvents(this.props['update_on'], self.update)
     this.registerEvents(this.props['replace_on'], self.replace)
   },
-  beforeDestroy: function() {
+  beforeUnmount: function() {
     const self = this
     this.cableTemplate = null
     this.removeEvents(this.props['append_on'], self.append)
@@ -89,12 +89,8 @@ const componentDef = {
     this.removeEvents(this.props['delete_on'], self.delete)
     this.removeEvents(this.props['update_on'], self.update)
     this.removeEvents(this.props['replace_on'], self.replace)
-  },
-  components: {
-    VRuntimeTemplate: VRuntimeTemplate
   }
 }
 
-let component = Vue.component('matestack-ui-core-cable', componentDef)
 
 export default componentDef
