@@ -1,10 +1,12 @@
-import Vue from 'vue'
 import axios from 'axios'
-import VRuntimeTemplate from "vue3-runtime-template"
 import matestackEventHub from '../event_hub'
+import componentMixin from './mixin'
+import componentHelpers from './helpers'
 
 const componentDef = {
+  mixins: [componentMixin],
   props: ['props', 'params'],
+  template: componentHelpers.inlineTemplate,
   data: function(){
     return {
       isolatedTemplate: null,
@@ -33,7 +35,7 @@ const componentDef = {
         method: "get",
         url: location.pathname + location.search,
         headers: {
-          'X-CSRF-Token': document.getElementsByName("csrf-token")[0].getAttribute('content')
+          'X-CSRF-Token': self.getXcsrfToken()
         },
         params: {
           "component_class": self.props["component_class"],
@@ -62,7 +64,7 @@ const componentDef = {
   created: function () {
     const self = this
   },
-  beforeDestroy: function() {
+  beforeUnmount: function() {
     const self = this
     if(this.props["rerender_on"] != undefined){
       var rerender_events = this.props["rerender_on"].split(",")
@@ -97,12 +99,7 @@ const componentDef = {
       rerender_events.forEach(rerender_event => matestackEventHub.$on(rerender_event.trim(), self.rerender));
     }
 
-  },
-  components: {
-    VRuntimeTemplate: VRuntimeTemplate
   }
 }
-
-let component = Vue.component('matestack-ui-core-isolate', componentDef)
 
 export default componentDef
