@@ -1,9 +1,10 @@
-import Vue from 'vue'
 import matestackEventHub from '../event_hub'
 import componentMixin from './mixin'
+import componentHelpers from './helpers'
 
 const componentDef = {
   mixins: [componentMixin],
+  template: componentHelpers.inlineTemplate,
   data: function(){
     return {
       showing: true,
@@ -50,7 +51,7 @@ const componentDef = {
       this.showing = true
     }
   },
-  beforeDestroy: function() {
+  beforeUnmount: function() {
     const self = this
     clearTimeout(self.hide_after_timeout)
     matestackEventHub.$off(this.props["show_on"], self.show);
@@ -65,7 +66,5 @@ const componentDef = {
     }
   },
 }
-
-let component = Vue.component('matestack-ui-core-toggle', componentDef)
 
 export default componentDef
