@@ -2,17 +2,30 @@ module Matestack
   module Ui
     module Core
       class App < Base
-        
+
         def initialize(options = {})
           @controller = Context.controller
           Context.app = self
           super(nil, nil, options)
         end
 
+        def create_children &block
+          self.app_wrapper do
+            self.response &block
+          end
+        end
+
+        def app_wrapper(&block)
+          Base.new(:component, component_attributes) do
+            div(class: 'matestack-app-wrapper', &block)
+          end
+        end
+
         def component_attributes
           {
             is: 'matestack-ui-core-app',
-            'inline-template': true,
+            ':params': params.to_json,
+            'v-slot': "{ vc, vueComponent }"
           }
         end
 
@@ -25,11 +38,11 @@ module Matestack
           subclass.layout(@layout)
           super
         end
-          
+
         def self.layout(layout = nil)
           @layout = layout ? layout : @layout
         end
-        
+
       end
     end
   end
