@@ -1,13 +1,25 @@
-import Vue from 'vue'
-import Vuex from 'vuex'
+import { inject } from 'vue'
+
 import axios from 'axios'
 import matestackEventHub from '../event_hub'
 import componentMixin from './mixin'
+import componentHelpers from './helpers'
+
+import transitionHandlingMixin from './transition_handling_mixin'
 
 const componentDef = {
-  mixins: [componentMixin],
+  mixins: [componentMixin, transitionHandlingMixin],
+  template: componentHelpers.inlineTemplate,
   data: function(){
     return {}
+  },
+  setup() {
+    // conditionally inject appNavigateTo
+    // action component has to work in context without wrapping app as well!
+    const appNavigateTo = inject('appNavigateTo', undefined)
+    return {
+      appNavigateTo
+    }
   },
   methods: {
     perform: function(){
@@ -35,7 +47,7 @@ const componentDef = {
           url: self.props["action_path"],
           data: self.props["data"],
           headers: {
-            'X-CSRF-Token': document.getElementsByName("csrf-token")[0].getAttribute('content')
+            'X-CSRF-Token': self.getXcsrfToken()
           }
         }
       )
@@ -43,109 +55,16 @@ const componentDef = {
         if (self.props["success"] != undefined && self.props["success"]["emit"] != undefined) {
           matestackEventHub.$emit(self.props["success"]["emit"], response.data);
         }
-
-        // transition handling
-        if (self.props["success"] != undefined
-          && self.props["success"]["transition"] != undefined
-          && (
-            self.props["success"]["transition"]["follow_response"] == undefined
-            ||
-            self.props["success"]["transition"]["follow_response"] === false
-          )
-          && self.$store != undefined
-        ) {
-          let path = self.props["success"]["transition"]["path"]
-          self.$store.dispatch('navigateTo', {url: path, backwards: false})
-          return;
-        }
-        if (self.props["success"] != undefined
-          && self.props["success"]["transition"] != undefined
-          && self.props["success"]["transition"]["follow_response"] === true
-          && self.$store != undefined
-        ) {
-          let path = response.data["transition_to"] || response.request.responseURL
-          self.$store.dispatch('navigateTo', {url: path, backwards: false})
-          return;
-        }
-        // redirect handling
-        if (self.props["success"] != undefined
-          && self.props["success"]["redirect"] != undefined
-          && (
-            self.props["success"]["redirect"]["follow_response"] == undefined
-            ||
-            self.props["success"]["redirect"]["follow_response"] === false
-          )
-          && self.$store != undefined
-        ) {
-          let path = self.props["success"]["redirect"]["path"]
-          window.location.href = path
-          return;
-        }
-        if (self.props["success"] != undefined
-          && self.props["success"]["redirect"] != undefined
-          && self.props["success"]["redirect"]["follow_response"] === true
-          && self.$store != undefined
-        ) {
-          let path = response.data["redirect_to"] || response.request.responseURL
-          window.location.href = path
-          return;
-        }
+        self.successTransitionHandling(response)
       })
       .catch(function(error){
         if (self.props["failure"] != undefined && self.props["failure"]["emit"] != undefined) {
           matestackEventHub.$emit(self.props["failure"]["emit"], error.response.data);
         }
-        // transition handling
-        if (self.props["failure"] != undefined
-          && self.props["failure"]["transition"] != undefined
-          && (
-            self.props["failure"]["transition"]["follow_response"] == undefined
-            ||
-            self.props["failure"]["transition"]["follow_response"] === false
-          )
-          && self.$store != undefined
-        ) {
-          let path = self.props["failure"]["transition"]["path"]
-          self.$store.dispatch('navigateTo', {url: path, backwards: false})
-          return;
-        }
-        if (self.props["failure"] != undefined
-          && self.props["failure"]["transition"] != undefined
-          && self.props["failure"]["transition"]["follow_response"] === true
-          && self.$store != undefined
-        ) {
-          let path = error.response.data["transition_to"] || response.request.responseURL
-          self.$store.dispatch('navigateTo', {url: path, backwards: false})
-          return;
-        }
-        // redirect handling
-        if (self.props["failure"] != undefined
-          && self.props["failure"]["redirect"] != undefined
-          && (
-            self.props["failure"]["redirect"]["follow_response"] == undefined
-            ||
-            self.props["failure"]["redirect"]["follow_response"] === false
-          )
-          && self.$store != undefined
-        ) {
-          let path = self.props["failure"]["redirect"]["path"]
-          window.location.href = path
-          return;
-        }
-        if (self.props["failure"] != undefined
-          && self.props["failure"]["redirect"] != undefined
-          && self.props["failure"]["redirect"]["follow_response"] === true
-          && self.$store != undefined
-        ) {
-          let path = error.response.data["redirect_to"] || response.request.responseURL
-          window.location.href = path
-          return;
-        }
+        self.failureTransitionHandling(error)
       })
     }
   }
 }
-
-let component = Vue.component('matestack-ui-core-action', componentDef)
 
 export default componentDef
