@@ -1,10 +1,11 @@
-import Vue from 'vue'
 import matestackEventHub from '../../event_hub'
 import queryParamsHelper from '../../helpers/query_params_helper'
 import componentMixin from '../mixin'
+import componentHelpers from '../helpers'
 
 const componentDef = {
   mixins: [componentMixin],
+  template: componentHelpers.inlineTemplate,
   data: function(){
     return {
       currentLimit: null,
@@ -91,7 +92,5 @@ const componentDef = {
     }
   }
 }
-
-let component = Vue.component('matestack-ui-core-collection-content', componentDef)
 
 export default componentDef
