@@ -1,11 +1,12 @@
-import Vue from 'vue'
 import matestackEventHub from '../../event_hub'
 import queryParamsHelper from '../../helpers/query_params_helper'
+import componentHelpers from '../helpers'
 
 import formMixin from '../form/form'
 
 const componentDef = {
   mixins: [formMixin],
+  template: componentHelpers.inlineTemplate,
   methods: {
     perform: function(){
       var url;
@@ -40,7 +41,5 @@ const componentDef = {
     })
   }
 }
-
-let component = Vue.component('matestack-ui-core-collection-filter', componentDef)
 
 export default componentDef
