@@ -4,13 +4,13 @@ module Matestack
       module Components
         module Collection
           class Next < Matestack::Ui::Component
-            
+
             def response
-              a options.merge('@click': 'next()') do
+              a options.merge('v-on:click': 'vc.next()') do
                 yield
               end
             end
-            
+
           end
         end
       end
