@@ -1,15 +1,13 @@
-import Vue from "vue";
-
 import formCheckboxMixin from "./checkbox_mixin";
 import componentMixin from "../mixin";
+import componentHelpers from "../helpers";
 
 const componentDef = {
   mixins: [componentMixin, formCheckboxMixin],
+  template: componentHelpers.inlineTemplate,
   data() {
     return {};
   }
 }
-
-let component = Vue.component("matestack-ui-core-form-checkbox", componentDef);
 
 export default componentDef;
