@@ -6,7 +6,7 @@ module Matestack
           class FieldsForRemoveItem < Matestack::Ui::Component
 
             def response
-              a class: 'matestack-ui-core-form-fields-for-remove-item', "@click.prevent": "removeItem()" do
+              a class: 'matestack-ui-core-form-fields-for-remove-item', "@click.prevent": "vc.removeItem()" do
                 yield if block_given?
               end
             end
