@@ -1,28 +1,51 @@
-import Vue from "vue";
-import Vuex from "vuex";
-import VRuntimeTemplate from "vue3-runtime-template"
-
+import { inject, computed } from "vue";
 import axios from "axios";
 
 import matestackEventHub from "../../event_hub";
 import componentMixin from "../mixin";
+import componentHelpers from "../helpers";
+
+import transitionHandlingMixin from '../transition_handling_mixin'
 
 const componentDef = {
-  mixins: [componentMixin],
+  mixins: [componentMixin, transitionHandlingMixin],
+  template: componentHelpers.inlineTemplate,
   data: function () {
     return {
       data: {},
       errors: {},
       loading: false,
-      nestedForms: {},
       isNestedForm: false,
-      hideNestedForm: false,
+      nestedForms: {},
       nestedFormRuntimeTemplates: {},
       nestedFormRuntimeTemplateDomElements: {},
-      deletedNestedForms: {},
-      nestedFormRuntimeId: "",
-      nestedFormServerErrorIndex: "",
+      deletedNestedForms: {}
     };
+  },
+  setup() {
+    // conditionally inject appNavigateTo
+    // form component has to work in context without wrapping app as well!
+    const appNavigateTo = inject('appNavigateTo', undefined)
+    return {
+      appNavigateTo
+    }
+  },
+  provide: function() {
+    return {
+      parentFormUid: computed(() => this.props["component_uid"]),
+      parentFormData: computed(() => this.data),
+      parentFormErrors: computed(() => this.errors),
+      parentFormLoading: computed(() => this.loading),
+      parentNestedForms: computed(() => this.nestedForms),
+      parentDeletedNestedForms: computed(() => this.deletedNestedForms),
+      parentNestedFormRuntimeTemplates: computed(() => this.nestedFormRuntimeTemplates),
+      parentNestedFormRuntimeTemplateDomElements: computed(() => this.nestedFormRuntimeTemplateDomElements),
+      parentFormMapToNestedForms: this.mapToNestedForms,
+      parentFormSetErrors: this.setErrors,
+      parentFormResetErrors: this.resetErrors,
+      parentFormIsNestedForm: this.isNestedForm, // need to provide this value for input components working both in form and nested form contexts
+      parentNestedFormRuntimeId: null, // need to provide this value for input components working both in form and nested form contexts
+    }
   },
   methods: {
     initDataKey: function (key, initValue) {
@@ -43,24 +66,13 @@ const componentDef = {
     resetErrors: function (key) {
       if (this.errors[key]) {
         delete this.errors[key];
-        Vue.set(this.errors);
-      }
-      if (this.isNestedForm){
-        var serverErrorKey = this.props["fields_for"].replace("_attributes", "")+"["+this.nestedFormServerErrorIndex+"]."+key
-        if (this.$parent.errors[serverErrorKey]) {
-          delete this.$parent.errors[serverErrorKey];
-          Vue.set(this.$parent.errors);
-        }
       }
     },
     setErrors: function(errors){
       this.errors = errors;
     },
-    setNestedFormServerErrorIndex: function(value){
-      this.nestedFormServerErrorIndex = value;
-    },
     setErrorKey: function(key, value){
-      Vue.set(this.errors, key, value);
+      this.errors[key] = value;
     },
     flushErrors: function(key, value){
       this.errors = {};
@@ -106,77 +118,16 @@ const componentDef = {
             var destroyed = true;
           }
           nestedFormInstance.initValues()
-          Vue.set(nestedFormInstance.data)
           if(destroyed){
             nestedFormInstance.hideNestedForm = true
-            Vue.set(nestedFormInstance.data, "_destroy", true)
+            nestedFormInstance.data["_destroy"] = true
           }
         })
       })
     },
-    removeItem: function(){
-      Vue.set(this.data, "_destroy", true)
-      this.hideNestedForm = true;
-      var id = parseInt(this.nestedFormRuntimeId.replace("_"+this.props["fields_for"]+"_child_", ""));
-      this.$parent.deletedNestedForms[this.props["fields_for"]].push(id);
-      var serverErrorKey = this.props["fields_for"].replace("_attributes", "")+"["+this.nestedFormServerErrorIndex+"]."
-      var self = this;
-      Object.keys(self.$parent.errors).forEach(function(errorKey){
-        if (errorKey.lastIndexOf(serverErrorKey, 0) == 0) {
-          delete self.$parent.errors[errorKey];
-          Vue.set(self.$parent.errors)
-        }
-      });
-    },
-    addItem: function(key){
-      var templateString = JSON.parse(this.$el.querySelector('#prototype-template-for-'+key).dataset[":template"])
-      if (this.nestedFormRuntimeTemplateDomElements[key] == null){
-        var dom_elem = document.createElement('div')
-        dom_elem.innerHTML = templateString
-        var existingItemsCount;
-        if (this.nestedForms[key] == undefined){
-          existingItemsCount = 0
-        }else{
-          existingItemsCount = this.nestedForms[key].length
-        }
-        dom_elem.querySelector('.matestack-form-fields-for').id = key+"_child_"+existingItemsCount
-        Vue.set(this.nestedFormRuntimeTemplateDomElements, key, dom_elem)
-        Vue.set(this.nestedFormRuntimeTemplates, key, this.nestedFormRuntimeTemplateDomElements[key].outerHTML)
-      }else{
-        var dom_elem = document.createElement('div')
-        dom_elem.innerHTML = templateString
-        var existingItemsCount = this.nestedForms[key].length
-        dom_elem.querySelector('.matestack-form-fields-for').id = key+"_child_"+existingItemsCount
-        this.nestedFormRuntimeTemplateDomElements[key].insertAdjacentHTML(
-          'beforeend',
-          dom_elem.innerHTML
-        )
-        Vue.set(this.nestedFormRuntimeTemplates, key, this.nestedFormRuntimeTemplateDomElements[key].outerHTML)
-      }
-    },
     initValues: function () {
-      let self = this;
-      let data = {};
-
-      for (let key in self.$refs) {
-        if (key.startsWith("input-component")) {
-          self.$refs[key].initialize()
-        }
-        if (key.startsWith("textarea-component")) {
-          self.$refs[key].initialize()
-        }
-        if (key.startsWith("select-component")) {
-          self.$refs[key].initialize()
-        }
-        if (key.startsWith("radio-component")) {
-          self.$refs[key].initialize()
-        }
-        if (key.startsWith("checkbox-component")) {
-          self.$refs[key].initialize()
-        }
-      }
+      this.emitScopedEvent("init") // received by child input components
     },
-
     shouldResetFormOnSuccessfulSubmit() {
       const self = this;
       if (self.props["success"] != undefined && self.props["success"]["reset"] != undefined) {
@@ -198,7 +149,9 @@ const componentDef = {
       if (self.props["fields_for"] != null) {
         return;
       }
-      var form = self.$el.tagName == 'FORM' ? self.$el : self.$el.querySelector('form');
+
+      var form = this.getRefs()["form"]
+
       if(form.checkValidity()){
         self.loading = true;
         if (self.props["emit"] != undefined) {
@@ -265,7 +218,7 @@ const componentDef = {
           url: self.props["submit_path"],
           data: formData,
           headers: {
-            "X-CSRF-Token": document.getElementsByName("csrf-token")[0].getAttribute("content"),
+            "X-CSRF-Token": self.getXcsrfToken(),
             "Content-Type": "multipart/form-data",
           },
         };
@@ -275,7 +228,7 @@ const componentDef = {
           url: self.props["submit_path"],
           data: payload,
           headers: {
-            "X-CSRF-Token": document.getElementsByName("csrf-token")[0].getAttribute("content"),
+            "X-CSRF-Token": self.getXcsrfToken(),
             "Content-Type": "application/json",
           },
         };
@@ -283,56 +236,12 @@ const componentDef = {
       axios(axios_config)
         .then(function (response) {
           self.loading = false;
+
           if (self.props["success"] != undefined && self.props["success"]["emit"] != undefined) {
             matestackEventHub.$emit(self.props["success"]["emit"], response.data);
           }
-          // transition handling
-          if (self.props["success"] != undefined
-            && self.props["success"]["transition"] != undefined
-            && (
-              self.props["success"]["transition"]["follow_response"] == undefined
-              ||
-              self.props["success"]["transition"]["follow_response"] === false
-            )
-            && self.$store != undefined
-          ) {
-            let path = self.props["success"]["transition"]["path"]
-            self.$store.dispatch('navigateTo', {url: path, backwards: false})
-            return;
-          }
-          if (self.props["success"] != undefined
-            && self.props["success"]["transition"] != undefined
-            && self.props["success"]["transition"]["follow_response"] === true
-            && self.$store != undefined
-          ) {
-            let path = response.data["transition_to"] || response.request.responseURL
-            self.$store.dispatch('navigateTo', {url: path, backwards: false})
-            return;
-          }
-          // redirect handling
-          if (self.props["success"] != undefined
-            && self.props["success"]["redirect"] != undefined
-            && (
-              self.props["success"]["redirect"]["follow_response"] == undefined
-              ||
-              self.props["success"]["redirect"]["follow_response"] === false
-            )
-            && self.$store != undefined
-          ) {
-            let path = self.props["success"]["redirect"]["path"]
-            window.location.href = path
-            return;
-          }
-          if (self.props["success"] != undefined
-            && self.props["success"]["redirect"] != undefined
-            && self.props["success"]["redirect"]["follow_response"] === true
-            && self.$store != undefined
-          ) {
-            let path = response.data["redirect_to"] || response.request.responseURL
-            window.location.href = path
-            return;
-          }
 
+          self.successTransitionHandling(response)
           self.flushErrors();
 
           if (self.shouldResetFormOnSuccessfulSubmit())
@@ -343,131 +252,25 @@ const componentDef = {
         })
         .catch(function (error) {
           self.loading = false;
+
           if (error.response && error.response.data && error.response.data.errors) {
             self.errors = error.response.data.errors;
             self.setErrors(error.response.data.errors);
             self.setNestedFormsError(error.response.data.errors);
           }
+
           if (self.props["failure"] != undefined && self.props["failure"]["emit"] != undefined) {
             matestackEventHub.$emit(self.props["failure"]["emit"], error.response.data);
           }
-          // transition handling
-          if (self.props["failure"] != undefined
-            && self.props["failure"]["transition"] != undefined
-            && (
-              self.props["failure"]["transition"]["follow_response"] == undefined
-              ||
-              self.props["failure"]["transition"]["follow_response"] === false
-            )
-            && self.$store != undefined
-          ) {
-            let path = self.props["failure"]["transition"]["path"]
-            self.$store.dispatch('navigateTo', {url: path, backwards: false})
-            return;
-          }
-          if (self.props["failure"] != undefined
-            && self.props["failure"]["transition"] != undefined
-            && self.props["failure"]["transition"]["follow_response"] === true
-            && self.$store != undefined
-          ) {
-            let path = error.response.data["transition_to"] || response.request.responseURL
-            self.$store.dispatch('navigateTo', {url: path, backwards: false})
-            return;
-          }
-          // redirect handling
-          if (self.props["failure"] != undefined
-            && self.props["failure"]["redirect"] != undefined
-            && (
-              self.props["failure"]["redirect"]["follow_response"] == undefined
-              ||
-              self.props["failure"]["redirect"]["follow_response"] === false
-            )
-            && self.$store != undefined
-          ) {
-            let path = self.props["failure"]["redirect"]["path"]
-            window.location.href = path
-            return;
-          }
-          if (self.props["failure"] != undefined
-            && self.props["failure"]["redirect"] != undefined
-            && self.props["failure"]["redirect"]["follow_response"] === true
-            && self.$store != undefined
-          ) {
-            let path = error.response.data["redirect_to"] || response.request.responseURL
-            window.location.href = path
-            return;
-          }
+
+          self.failureTransitionHandling(error)
         });
     },
   },
   mounted: function () {
-    var self = this;
-    if (this.props["fields_for"] != undefined) {
-      this.isNestedForm = true;
-
-      this.data = { "_destroy": false };
-
-      //initialize nestedForm data in parent form if required
-      if(this.$parent.data[this.props["fields_for"]] == undefined){
-        this.$set(this.$parent.data, this.props["fields_for"], []);
-      }
-      if(this.$parent.nestedForms[this.props["fields_for"]] == undefined){
-        this.$set(this.$parent.nestedForms, this.props["fields_for"], []);
-      }
-      if(this.$parent.deletedNestedForms[this.props["fields_for"]] == undefined){
-        this.$set(this.$parent.deletedNestedForms, this.props["fields_for"], []);
-      }
-
-      var id = parseInt(self.$el.id.replace(this.props["fields_for"]+"_child_", ""));
-
-      //setup data binding for serverside rendered nested forms
-      if (isNaN(id)){
-        id = this.$parent.nestedForms[this.props["fields_for"]].length
-        this.nestedFormRuntimeId = "_"+this.props["fields_for"]+"_child_"+id
-        this.$el.id = this.props["fields_for"]+"_child_"+id
-        this.initValues()
-        this.$parent.data[this.props["fields_for"]].push(this.data);
-        this.$parent.nestedForms[this.props["fields_for"]].push(this);
-      }
-
-      //setup data binding for runtime nested forms (dynamic add via v-runtime-template)
-      if (!isNaN(id)){
-        this.nestedFormRuntimeId = "_"+this.props["fields_for"]+"_child_"+id
-        if(this.$parent.data[this.props["fields_for"]][id] == undefined){
-          //new runtime form
-          this.initValues()
-          this.$parent.data[this.props["fields_for"]].push(this.data);
-          this.$parent.nestedForms[this.props["fields_for"]].push(this);
-        }else{
-          //retreive state for existing runtime form (after remount for example)
-          this.data = this.$parent.data[this.props["fields_for"]][id]
-          if (this.data["_destroy"] == true){
-            this.hideNestedForm = true;
-          }
-          this.$parent.nestedForms[this.props["fields_for"]][id] = this;
-          Object.keys(this.$parent.errors).forEach(function(errorKey){
-            if (errorKey.includes(".")){
-              let childErrorKey = errorKey.split(".")[1]
-              let childModelName = errorKey.split(".")[0].split("[")[0]
-              let childModelIndex = errorKey.split(".")[0].split("[")[1].split("]")[0]
-              let mappedChildModelIndex = self.$parent.mapToNestedForms(parseInt(childModelIndex), childModelName+"_attributes")
-              if(childModelName+"_attributes" == self.props["fields_for"] && mappedChildModelIndex == id){
-                self.setNestedFormServerErrorIndex(parseInt(childModelIndex))
-                self.setErrorKey(childErrorKey, self.$parent.errors[errorKey])
-              }
-            }
-          })
-        }
-      }
-    } else {
-      this.initValues();
-    }
-  },
-  components: {
-    VRuntimeTemplate: VRuntimeTemplate
+    this.initValues();
   }
-};
 
-let component = Vue.component("matestack-ui-core-form", componentDef);
+};
 
 export default componentDef;
