@@ -7,9 +7,6 @@ module Matestack
             vue_name 'matestack-ui-core-form'
 
             optional :for, :path, :success, :failure, :multipart, :emit, :delay, :errors
-            optional :fields_for, :reject_blank
-
-            attr_accessor :prototype_template
 
             # setup form context to allow child components like inputs to access the form configuration
             def initialize(html_tag = nil, text = nil, options = {}, &block)
@@ -19,29 +16,18 @@ module Matestack
               Matestack::Ui::VueJs::Components::Form::Context.form_context = previous_form_context
             end
 
-            def component_id
-              "matestack-form-fields-for-#{context.fields_for}-#{SecureRandom.hex}" if context.fields_for
-            end
-
             def response
-              if context.fields_for
-                div class: "matestack-form-fields-for", "v-show": "hideNestedForm != true", id: options[:id] do
-                  form_input key: context.for&.class&.primary_key, type: :hidden # required for existing model mapping
-                  form_input key: :_destroy, type: :hidden, init: true if context.reject_blank == true
-                  yield
-                end
-              else
-                form attributes do
-                  yield
-                end
+              form attributes do
+                yield
               end
             end
 
             def attributes
               {
                 class: 'matestack-form',
-                'v-bind:class': "{ 'has-errors': hasErrors(), loading: loading }",
-                '@submit.prevent': 'perform',
+                "matestack-ui-core-ref": scoped_ref('form'),
+                'v-bind:class': "{ 'has-errors': vc.hasErrors(), loading: vc.loading }",
+                'v-on:submit.prevent': 'vc.perform',
               }
             end
 
@@ -54,9 +40,7 @@ module Matestack
                 failure: ctx.failure,
                 multipart: !!ctx.multipart,
                 emit: ctx.emit,
-                delay: ctx.delay,
-                fields_for: ctx.fields_for,
-                primary_key: for_object_primary_key
+                delay: ctx.delay
               }
             end
 
@@ -69,6 +53,10 @@ module Matestack
               @for_option ||= ctx.for
             end
 
+            def multipart_option
+              @multipart_option ||= ctx.multipart
+            end
+
             def for_object_primary_key
               context.for&.class&.primary_key rescue nil
             end
@@ -76,6 +64,11 @@ module Matestack
             def form_method
               @form_method ||= options.delete(:method)
             end
+
+            def is_nested_form?
+              false
+            end
+
           end
         end
       end
