@@ -1,15 +1,13 @@
-import Vue from "vue";
-
 import formInputMixin from "./input_mixin";
 import componentMixin from "../mixin";
+import componentHelpers from "../helpers";
 
 const componentDef = {
   mixins: [componentMixin, formInputMixin],
+  template: componentHelpers.inlineTemplate,
   data() {
     return {};
   }
 }
-
-let component = Vue.component("matestack-ui-core-form-input", componentDef);
 
 export default componentDef;
