@@ -1,20 +1,33 @@
+import matestackEventHub from '../../event_hub'
+
 const formInputMixin = {
+  inject: [
+    'parentFormUid',
+    'parentFormData',
+    'parentFormErrors',
+    'parentFormLoading',
+    'parentFormIsNestedForm',
+    'parentFormResetErrors',
+    'parentNestedFormRuntimeId'
+  ],
   methods: {
     initialize: function(){
       const self = this
       let data = {};
 
-      for (let key in this.$refs) {
-        let initValue = this.$refs[key]["attributes"]["init-value"];
+      for (let key in self.getRefs()) {
+        if (key.startsWith("input.")) {
+          let initValue = self.getRefs()[key]["attributes"]["init-value"];
 
-        self.$set(self.$parent.data, key.replace("input.", ""), null)
+          self.parentFormData[key.replace("input.", "")] = null
 
-        if (initValue) {
-          self.setValue(initValue["value"])
-          self.afterInitialize(initValue["value"])
-        } else {
-          self.setValue(null)
-          self.afterInitialize(null)
+          if (initValue) {
+            self.setValue(initValue["value"])
+            self.afterInitialize(initValue["value"])
+          } else {
+            self.setValue(null)
+            self.afterInitialize(null)
+          }
         }
       }
 
@@ -23,30 +36,43 @@ const formInputMixin = {
       const dataTransfer = event.dataTransfer || event.target;
       const files = dataTransfer.files;
       if (event.target.attributes.multiple) {
-        this.$parent.data[key] = [];
+        this.parentFormData[key] = [];
         for (let index in files) {
           if (files[index] instanceof File) {
-            this.$parent.data[key].push(files[index]);
+            this.parentFormData[key].push(files[index]);
           }
         }
       } else {
-        this.$parent.data[key] = files[0];
+        this.parentFormData[key] = files[0];
       }
     },
     inputChanged: function (key) {
-      if (this.$parent.isNestedForm){
-        this.$parent.data["_destroy"] = false;
+      if (this.parentFormIsNestedForm){
+        this.parentFormData["_destroy"] = false;
       }
-      this.$parent.resetErrors(key);
+      this.parentFormResetErrors(key);
 
     },
     afterInitialize: function(value){
       // can be used in the main component for further initialization steps
     },
     setValue: function (value){
-      this.$parent.data[this.props["key"]] = value;
+      if(this.getRefs()["input."+this.props["key"]]){
+        this.getRefs()["input."+this.props["key"]].value = value;  // reset file upload inputs properly
+      }
+      if(this.getRefs()["input."+this.props["key"]+"[]"]){
+        this.getRefs()["input."+this.props["key"]+"[]"].value = value;  // reset multiple file upload inputs properly
+      }
+      this.parentFormData[this.props["key"]] = value;
     }
+  },
+  mounted: function(){
+    this.registerScopedEvent("init", this.initialize, this.parentFormUid)
+  },
+  beforeUnmount: function(){
+    this.removeScopedEvent("init", this.initialize, this.parentFormUid)
   }
+
 
 }
 
