@@ -1,16 +1,30 @@
 const formRadioMixin = {
+  inject: [
+    'parentFormUid',
+    'parentFormData',
+    'parentFormErrors',
+    'parentFormLoading',
+    'parentFormIsNestedForm',
+    'parentFormResetErrors',
+    'parentNestedFormRuntimeId'
+  ],
   methods: {
     initialize: function(){
       const self = this
       let data = {};
 
-      for (let key in self.$refs) {
-        let initValue = self.$refs[key]["attributes"]["init-value"];
-        let valueType = self.$refs[key]["attributes"]["value-type"];
+      for (let key in self.getRefs()) {
+        let initValue;
+        let valueType;
+
+        if (key.startsWith("select.")) {
+          initValue = self.getRefs()[key]["attributes"]["init-value"];
+          valueType = self.getRefs()[key]["attributes"]["value-type"];
+        }
 
         if (key.startsWith("select.")) {
           if (key.startsWith("select.multiple.")) {
-            self.$set(self.$parent.data, key.replace("select.multiple.", ""), null)
+            self.parentFormData[key.replace("select.multiple.", "")] = null
             if (initValue) {
               self.setValue(JSON.parse(initValue["value"]));
               self.afterInitialize(JSON.parse(initValue["value"]))
@@ -19,7 +33,7 @@ const formRadioMixin = {
               self.afterInitialize([]);
             }
           } else {
-            self.$set(self.$parent.data, key.replace("select.", ""), null)
+            self.parentFormData[key.replace("select.", "")] = null
             if (initValue) {
               if (valueType && valueType["value"] == "Integer") {
                 self.setValue(parseInt(initValue["value"]));
@@ -37,17 +51,23 @@ const formRadioMixin = {
       }
     },
     inputChanged: function (key) {
-      if (this.$parent.isNestedForm){
-        this.$parent.data["_destroy"] = false;
+      if (this.parentFormIsNestedForm){
+        this.parentFormData["_destroy"] = false;
       }
-      this.$parent.resetErrors(key);
+      this.parentFormResetErrors(key);
     },
     afterInitialize: function(value){
       // can be used in the main component for further initialization steps
     },
     setValue: function (value){
-      this.$parent.data[this.props["key"]] = value
+      this.parentFormData[this.props["key"]] = value
     }
+  },
+  mounted: function(){
+    this.registerScopedEvent("init", this.initialize, this.parentFormUid)
+  },
+  beforeUnmount: function(){
+    this.removeScopedEvent("init", this.initialize, this.parentFormUid)
   }
 
 }
