@@ -1,9 +1,10 @@
-import Vue from 'vue'
 import matestackEventHub from '../event_hub'
 import componentMixin from './mixin'
+import componentHelpers from './helpers'
 
 const componentDef = {
   mixins: [componentMixin],
+  template: componentHelpers.inlineTemplate,
   data: function(){
     return { }
   },
@@ -13,7 +14,5 @@ const componentDef = {
     }
   }
 }
-
-let component = Vue.component('matestack-ui-core-onclick', componentDef)
 
 export default componentDef
