@@ -1,26 +1,32 @@
-import Vue from 'vue'
-import Vuex from 'vuex'
-import componentMixin from './mixin'
 import matestackEventHub from '../event_hub'
+import componentMixin from './mixin'
+import componentHelpers from './helpers'
 
 const componentDef = {
   mixins: [componentMixin],
+  template: componentHelpers.inlineTemplate,
   data: function(){
     return {}
   },
-  computed: Vuex.mapState({
-    isActive (state) {
-      return (this.props["link_path"].split("?")[0]) === state.currentPathName
+  inject: [
+    'appNavigateTo',
+    'currentPathName',
+    'setPageLoading',
+    'pageLoading'
+  ],
+  computed: {
+    isActive () {
+      return (this.props["link_path"].split("?")[0]) === this.currentPathName
     },
-    isChildActive (state) {
-      return ((this.props["link_path"].split("?")[0]) !== state.currentPathName) && (state.currentPathName.indexOf(this.props["link_path"].split("?")[0]) !== -1)
+    isChildActive () {
+      return ((this.props["link_path"].split("?")[0]) !== this.currentPathName) && (this.currentPathName.indexOf(this.props["link_path"].split("?")[0]) !== -1)
     }
-  }),
+  },
   methods: {
     navigateTo: function(url){
       const self = this
       matestackEventHub.$emit("page_loading_triggered", url);
-      this.$store.commit('setPageLoading', true);
+      this.setPageLoading(true)
       if (self.props["delay"] != undefined) {
         setTimeout(function () {
           self.performNavigation(url)
@@ -30,13 +36,9 @@ const componentDef = {
       }
     },
     performNavigation: function(url){
-      this.$store.dispatch('navigateTo', {url: url, backwards: false}).then((response) => {
-        // self.asyncTemplate = response;
-      })
+      this.appNavigateTo({url: url, backwards: false})
     }
   }
 }
-
-let component = Vue.component('matestack-ui-core-transition', componentDef)
 
 export default componentDef
