@@ -1,5 +1,12 @@
-import Vue from 'vue'
+import mitt from 'mitt'
 
-const eventHub = new Vue();
+const emitter = mitt()
+
+const eventHub = {
+  $on: (...args) => emitter.on(...args),
+  $once: (...args) => emitter.once(...args),
+  $off: (...args) => emitter.off(...args),
+  $emit: (...args) => emitter.emit(...args)
+}
 
 export default eventHub
