@@ -1,23 +1,24 @@
-import Vue from 'vue'
-import Vuex from 'vuex'
 import componentMixin from '../components/mixin'
-
-import VRuntimeTemplate from "vue3-runtime-template"
+import componentHelpers from '../components/helpers'
 
 const componentDef = {
   mixins: [componentMixin],
+  template: componentHelpers.inlineTemplate,
   data: function(){
     return {}
   },
-  computed: Vuex.mapState({
-    asyncPageTemplate: state => state.pageTemplate,
-    loading: state => state.pageLoading
-  }),
-  components: {
-    VRuntimeTemplate: VRuntimeTemplate
+  inject: [
+    'pageTemplate',
+    'pageLoading'
+  ],
+  computed: {
+    asyncPageTemplate: function(){
+      return this.pageTemplate
+    },
+    loading: function(){
+      return this.pageLoading
+    }
   }
 }
-
-let component = Vue.component('matestack-ui-core-page-content', componentDef)
 
 export default componentDef
